functions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.7.2
azure-identity==1.7.1
azure-keyvault-secrets==4.3.0
azure-mgmt-datafactory==2.2.0
azure-storage-blob==12.9.0
azure-storage-file-datalake==12.5.0
beautifulsoup4==4.10.0
croniter==1.1.0
databricks-cli==0.16.2
O365==2.0.16
PyYAML==6.0
zipfile-deflate64==0.2.0